% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{centeringFeFunc}
\alias{centeringFeFunc}
\title{Performs feature engineering that can subtract values enabling centering}
\usage{
centeringFeFunc(trainData, centers)
}
\arguments{
\item{trainData}{The data to perform the feature engineering to}

\item{centers}{a data.frame with columns covariateId and centerValue}
}
\value{
The data with centering performed
}
\description{
Performs feature engineering that can subtract values enabling centering
}
\details{
This takes the featurEengineering settings for centering and does it
}
