% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{convertMeasurementFunc}
\alias{convertMeasurementFunc}
\title{Performs feature engineering for measurement mapping}
\usage{
convertMeasurementFunc(trainData, analysisId, covariateId, mappings)
}
\arguments{
\item{trainData}{The data to perform the feature engineering to description}

\item{analysisId}{the analysisId of the new covariates created by mapping the measurement}

\item{covariateId}{the covariateId of the measurementId to map}

\item{mappings}{A list of functions or string names of function to apply to the measurement}
}
\value{
The data with measurements mapped as specified
}
\description{
Performs feature engineering for measurement mapping
}
\details{
Given the settings this will do the measurement mapping as requested
}
