% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{createCenteringFe}
\alias{createCenteringFe}
\title{Feature engineering that can subtract values enabling centering}
\usage{
createCenteringFe(centers)
}
\arguments{
\item{centers}{a data.frame with columns covariateId and centerValue}
}
\value{
The feature engineering settings that can be used in runPlp to perform centering
}
\description{
Feature engineering that can subtract values enabling centering
}
\details{
The user specifies a data.frame with columns covariateId and centerValue if the covariateIds
are found in the data their value will be modified by subtracting the centerValue
}
