% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_measurement.R
\name{createCohortMeasurementCovariateSettings}
\alias{createCohortMeasurementCovariateSettings}
\title{Create covariates settings for measurements}
\usage{
createCohortMeasurementCovariateSettings(
  covariateName,
  cohortDatabaseSchema,
  cohortTable,
  cohortId,
  cohortStartDay,
  cohortEndDay,
  measurementConceptSet,
  measurementUnitSet = NULL,
  measurementStartDay = -30,
  measurementEndDay = 0,
  measurementScaleMap = function(x) {
     sapply(x, function(y) {
         y
     })
 },
  measurementMinVal = NULL,
  measurementMaxVal = NULL,
  measurementAggregateMethod = "recent",
  covariateId = 1566,
  analysisId = 566
)
}
\arguments{
\item{covariateName}{The name of the covariate}

\item{cohortDatabaseSchema}{The schema where the cohort is in}

\item{cohortTable}{The table name where the cohort is in}

\item{cohortId}{The cohortDefinitionId of the cohort to restrict to}

\item{cohortStartDay}{the start time relative to index to look for the person being in the cohort}

\item{cohortEndDay}{the end time relative to index to look for the person being in the cohort}

\item{measurementConceptSet}{A vector of conceptIds that are the measurement concepts}

\item{measurementUnitSet}{NULL or a vector of conceptIds to restrict the units to (can include NA)}

\item{measurementStartDay}{the start time before index to look for the measurement}

\item{measurementEndDay}{the end time before index to look for the measurement}

\item{measurementScaleMap}{A function that lets you concept units into a standard unit and do any scaling}

\item{measurementMinVal}{NULL or the min valid value for the measurements (value less than this are excluded)}

\item{measurementMaxVal}{NULL or the max valid value for the measurements (value more than this are excluded)}

\item{measurementAggregateMethod}{one of max/min/mean/median/recent how to handle multiple measurements}

\item{covariateId}{a unique value for the covariateId}

\item{analysisId}{a unique value for the analysisId}
}
\value{
An object of class `covariateSettings` specifying how to create the cohort covariate with the covariateId
 cohortId x 100000 + settingId x 1000 + analysisId
}
\description{
Create covariates settings for measurements
}
\details{
The user specifies
}
