% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{createInteractionFe}
\alias{createInteractionFe}
\title{Feature engineering that can add interaction terms for any covariate}
\usage{
createInteractionFe(
  interactionCovariateId,
  covariateIdsOfInterest,
  analysisId = 70
)
}
\arguments{
\item{interactionCovariateId}{the main covariateId such as the id for age in years}

\item{covariateIdsOfInterest}{a single covariateId or vector of covariateIds to interact with}

\item{analysisId}{A unique analysisId for the new interaction covariates}
}
\value{
The feature engineering settings that can be used in runPlp to add the interaction covariates
}
\description{
Feature engineering that can add interaction terms for any covariate
}
\details{
The user specifies a main covariateId called interactionCovariateId and other covariateIds to
interact with covariateIdsOfInterest and the code will create new covariateIds corresponding
to the interactions with the id of covariateIdsOfInterest*1000+analysisId
}
