% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_age.R
\name{createLogAgeCovariateSettings}
\alias{createLogAgeCovariateSettings}
\title{Create covariates settings for log age covariate}
\usage{
createLogAgeCovariateSettings(ageMultiply = 1)
}
\arguments{
\item{ageMultiply}{a value to multiple age by before logging}
}
\value{
An object of class `covariateSettings` specifying how to create the cohort covariate with the covariateId
 cohortId x 100000 + settingId x 1000 + analysisId
}
\description{
Create covariates settings for log age covariate
}
\details{
The user specifies
}
