% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement_covariates.R
\name{createMeasurementCovariateSettings}
\alias{createMeasurementCovariateSettings}
\title{Create covariates settings for measurements}
\usage{
createMeasurementCovariateSettings(
  covariateName,
  conceptSet,
  unitSet = NULL,
  startDay = -30,
  endDay = 0,
  scaleMap = function(x) {
     sapply(x, function(y) {
         y
     })
 },
  minVal = NULL,
  maxVal = NULL,
  aggregateMethod = "recent",
  covariateId = 1466,
  analysisId = 466
)
}
\arguments{
\item{covariateName}{The name of the covariate}

\item{conceptSet}{A vector of conceptIds that are the measurement concepts}

\item{unitSet}{NULL or a vector of conceptIds to restrict the units to (can include NA)}

\item{startDay}{the start time before index to look for the measurement}

\item{endDay}{the end time before index to look for the measurement}

\item{scaleMap}{A function that lets you concept units into a standard unit and do any scaling}

\item{minVal}{NULL or the min valid value for the measurements (value less than this are excluded)}

\item{maxVal}{NULL or the max valid value for the measurements (value more than this are excluded)}

\item{aggregateMethod}{one of max/min/mean/median/recent how to handle multiple measurements}

\item{covariateId}{a unique value for the covariateId}

\item{analysisId}{a unique value for the analysisId}
}
\value{
An object of class `covariateSettings` specifying how to create the cohort covariate with the covariateId
 cohortId x 100000 + settingId x 1000 + analysisId
}
\description{
Create covariates settings for measurements
}
\details{
The user specifies
}
