% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{createMeasurementFe}
\alias{createMeasurementFe}
\title{Feature engineering that can do mapping of a measurement variable}
\usage{
createMeasurementFe(
  covariateId = 2466,
  analysisId = 366,
  mappings = list(function(x) {
     sapply(x, function(x) 1/(x/10)^2)
 }, function(x) {

        sapply(x, function(x) (1/(x/10)^2) * log(x/10))
 })
)
}
\arguments{
\item{covariateId}{the covariateId of the measurementId to map}

\item{analysisId}{the analysisId of the new covariates created by mapping the measurement}

\item{mappings}{A list of functions or string names of function to apply to the measurement}
}
\value{
The feature engineering settings that can be used in runPlp to map the measurement
}
\description{
Feature engineering that can do mapping of a measurement variable
}
\details{
The user specifies the covariateId of the measurement covariate, an analysisId to use for the
new variables and a list of mappings that are performed the first mapping will have the new
covariateId of 1x1000+analysisId, the second mapping will have the new covariateId of 
2x1000+analysisId, etc.
}
