% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement_ratio_covariates.R
\name{createMeasurementRatioCovariateSettings}
\alias{createMeasurementRatioCovariateSettings}
\title{Create covariates settings for measurements}
\usage{
createMeasurementRatioCovariateSettings(
  covariateName,
  conceptSet1,
  conceptSet2,
  unitSet1 = NULL,
  unitSet2 = NULL,
  startDay = -30,
  endDay = 0,
  scaleMap1 = function(x) {
     return(x)
 },
  scaleMap2 = function(x) {
     return(x)
 },
  centeringMap = function(x) {
     return(x)
 },
  ageInteract = FALSE,
  minVal1 = NULL,
  maxVal1 = NULL,
  minVal2 = NULL,
  maxVal2 = NULL,
  aggregateMethod = "recent",
  covariateId = 1486,
  analysisId = 486
)
}
\arguments{
\item{covariateName}{The name of the covariate}

\item{conceptSet1}{A vector of conceptIds that are the numerator measurement concepts#'}

\item{conceptSet2}{A vector of conceptIds that are the denominator measurement concepts}

\item{unitSet1}{NULL or a vector of numerator conceptIds to restrict the units to (can include NA)}

\item{unitSet2}{NULL or a vector of denominator conceptIds to restrict the units to (can include NA)}

\item{startDay}{the start time before index to look for the measurement}

\item{endDay}{the end time before index to look for the measurement}

\item{scaleMap1}{A function that lets you map numerator concept units into a standard unit and do any scaling}

\item{scaleMap2}{A function that lets you map denominator concept units into a standard unit and do any scaling}

\item{centeringMap}{A function to center the final ratio value}

\item{ageInteract}{Whether to do interaction with age in years}

\item{minVal1}{NULL or the min valid value for the numerator measurements (value less than this are excluded)}

\item{maxVal1}{NULL or the max valid value for the numerator measurements (value more than this are excluded)}

\item{minVal2}{NULL or the min valid value for the denominator measurements (value less than this are excluded)}

\item{maxVal2}{NULL or the max valid value for the denominator measurements (value more than this are excluded)}

\item{aggregateMethod}{one of max/min/mean/median/recent how to handle multiple measurements}

\item{covariateId}{a unique value for the covariateId}

\item{analysisId}{a unique value for the analysisId}
}
\value{
An object of class `covariateSettings` specifying how to create the cohort covariate with the covariateId
 cohortId x 100000 + settingId x 1000 + analysisId
}
\description{
Create covariates settings for measurements
}
\details{
The user specifies
}
