
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{createPreferValueAndFillFe}
\alias{createPreferValueAndFillFe}
\title{Selects preferred covariate values and fill fallback covariate rows}
\usage{
createPreferValueAndFillFe(trainData, fallbackCovariateId, preferredCovariateId, dropPreferred = TRUE)
}
\arguments{
\item{trainData}{The data to perform the feature engineering to description}

\item{fallbackCovariateId}{The covariateId whose values should be overwritten or created}

\item{preferredCovariateId}{The covariateId whose values should be used when available}

\item{dropPreferred}{If TRUE, removes rows with \code{preferredCovariateId} after replacement}
}
\value{
The data with measurements mapped as specified
}
\description{
Selects preferred covariate values and fills fallback covariate row
}
\details{
Given the settings this will select a preferred value
}
