% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRisk1_male.R
\name{createQRISK1MaleOG10}
\alias{createQRISK1MaleOG10}
\title{Create qrisk 1 model for males usign Qrisk cohorts}
\usage{
createQRISK1MaleOG10(cohortDatabaseSchema, cohortTableName = "qrisk_cprd")
}
\arguments{
\item{cohortDatabaseSchema}{the schema where the cohorts are generated into}

\item{cohortTableName}{the cohort table name}
}
\value{
An plpModel
}
\description{
Create qrisk 1 model for males usign Qrisk cohorts
}
\details{
The user specifies
}
