% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRisk2A_male.R
\name{createQRISK2AMaleOG10}
\alias{createQRISK2AMaleOG10}
\title{Create qrisk 2A model for males using Qrisk cohorts}
\usage{
createQRISK2AMaleOG10(
  cohortDatabaseSchema,
  cohortTableName = "qrisk_cprd",
  sampleSize = 10000
)
}
\arguments{
\item{cohortDatabaseSchema}{the schema where the cohorts are generated into}

\item{cohortTableName}{the cohort table name}

\item{sampleSize}{the sample size to take of the target cohort if it is too large}
}
\value{
An plpModel
}
\description{
Create qrisk 2A model for males using Qrisk cohorts
}
\details{
The user specifies
}
