% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRisk2_male.R
\name{createQRISK2MaleOG10}
\alias{createQRISK2MaleOG10}
\title{Create qrisk 2 model for males using Qrisk cohorts}
\usage{
createQRISK2MaleOG10(cohortDatabaseSchema, cohortTableName = "qrisk_cprd")
}
\arguments{
\item{cohortDatabaseSchema}{the schema where the cohorts are generated into}

\item{cohortTableName}{the cohort table name}
}
\value{
An plpModel
}
\description{
Create qrisk 2 model for males using Qrisk cohorts
}
\details{
The user specifies
}
