% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRisk3_female.R
\name{createQRISK3FemaleOG10}
\alias{createQRISK3FemaleOG10}
\title{Create qrisk 3 model for females using Qrisk cohorts}
\usage{
createQRISK3FemaleOG10(cohortDatabaseSchema, cohortTableName = "qrisk_cprd")
}
\arguments{
\item{cohortDatabaseSchema}{the schema where the cohorts are generated into}

\item{cohortTableName}{the cohort table name}
}
\value{
An plpModel
}
\description{
Create qrisk 3 model for females using Qrisk cohorts
}
\details{
The user specifies
}
