% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRisk4_male.R
\name{createQRISK4MaleOG10}
\alias{createQRISK4MaleOG10}
\title{Create qrisk 4 model for males using Qrisk cohorts}
\usage{
createQRISK4MaleOG10(cohortDatabaseSchema, cohortTableName = "qrisk_cprd")
}
\arguments{
\item{cohortDatabaseSchema}{the schema where the cohorts are generated into}

\item{cohortTableName}{the cohort table name}
}
\value{
An plpModel
}
\description{
Create qrisk 4 model for males using Qrisk cohorts
}
\details{
The user specifies
}
