% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{interactionFeFunc}
\alias{interactionFeFunc}
\title{Performs feature engineering for interaction terms for any covariate}
\usage{
interactionFeFunc(
  trainData,
  interactionCovariateId,
  covariateIdsOfInterest,
  analysisId
)
}
\arguments{
\item{trainData}{The data to perform the feature engineering to description}

\item{interactionCovariateId}{the main covariateId such as the id for age in years}

\item{covariateIdsOfInterest}{a single covariateId or vector of covariateIds to interact with}

\item{analysisId}{A unique analysisId for the new interaction covariates}
}
\value{
The data with interactions as specified added
}
\description{
Performs feature engineering for interaction terms for any covariate
}
\details{
Given the settings this will do the interactions as requested
}
