% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRISKvalidation.R
\docType{package}
\name{QRISKvalidation}
\alias{QRISKvalidation-package}
\alias{QRISKvalidation}
\title{QRISKvalidation}
\description{
A package for running qrisk temporal validation
}
\author{
\strong{Maintainer}: Alexander Saelmans \email{Saelmans@ohdsi.org}

}
\keyword{internal}
